[SM-142] Rename CommCore to Commercial.Core (#2158)
* Rename CommCore to Commercial.Core
