+++ .github/workflows/build.yml
@@ -76,7 +76,7 @@ jobs:
         shell: pwsh
 
       - name: Test Bitwarden solution
-        run: dotnet test ./bitwarden_license/test/CmmCore.Test --configuration Debug --no-build
+        run: dotnet test ./bitwarden_license/test/Commercial.Core.Test --configuration Debug --no-build
         shell: pwsh
