+++ bitwarden_license/src/Commercial.Core/Commercial.Core.csproj
