+++ bitwarden_license/test/Commercial.Core.Test/Commercial.Core.Test.csproj
@@ -18,7 +18,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\CommCore\CommCore.csproj" />
+    <ProjectReference Include="..\..\src\Commercial.Core\Commercial.Core.csproj" />
     <ProjectReference Include="..\..\..\test\Core.Test\Core.Test.csproj" />
   </ItemGroup>
